pyproject.toml
[tool.poetry]
name = "de-zoomcamp-project"
version = "0.1.0"
description = ""
authors = ["wtfzambo <zambelli.federico@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.0"
pmaw = "^3.0.0"
python-dotenv = "^1.0.0"
duckdb = "0.7.1"
prefect = "^2.10.4"
asyncpraw = "^7.7.0"
praw = "^7.7.0"
prefect-gcp = {extras = ["bigquery"], version = "^0.4.1"}
pandas-gbq = "^0.19.1"


[tool.poetry.group.dev.dependencies]
notebook = "^6.5.4"
flake8 = "^6.0.0"
black = "^23.3.0"
flake8-black = "^0.3.6"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
force_single_line = false
lines_after_imports = 2
multi_line_output = 3